# Add json parsing performance tests and code clean up