--- a/src/System.Text.JsonLab/System/Text/Json/JsonParseObject.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonParseObject.cs
@@ -12,16 +12,34 @@ namespace System.Text.JsonLab
 {
     public ref struct JsonObject
     {
-        private MemoryPool<byte> _pool;
-        private IMemoryOwner<byte> _dbMemory;
-        private ReadOnlySpan<byte> _db; 
+        private Span<byte> _db; 
         private ReadOnlySpan<byte> _values;
 
-        public static JsonObject Parse(ReadOnlySpan<byte> utf8Json)
+        public string PrintDatabase()
         {
-            var parser = new JsonParser();
-            var result = parser.Parse(utf8Json);
-            return result;
+            StringBuilder sb = new StringBuilder();
+            sb.Append(nameof(Record.Location) + "\t" + nameof(Record.Length) + "\t" + nameof(Record.Type) + "\r\n");
+            for (int i = 0; i < _db.Length; i += DbRow.Size)
+            {
+                DbRow record = Read<DbRow>(_db.Slice(i));
+                sb.Append(record.Location + "\t" + record.Length + "\t" + record.Type + "\r\n");
+            }
+            return sb.ToString();
+        }
+
+        public string PrintJson()
+        {
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < _db.Length; i += DbRow.Size)
+            {
+                DbRow record = Read<DbRow>(_db.Slice(i));
+                if (record.IsSimpleValue)
+                {
+                    ReadOnlySpan<byte> value = _values.Slice(record.Location, record.Length);
+                    sb.Append(Encoding.UTF8.GetString(value.ToArray())).Append(", ");
+                }
+            }
+            return sb.ToString();
         }
 
         public static JsonObject Parse(ReadOnlySpan<byte> utf8Json, MemoryPool<byte> pool = null)
@@ -31,23 +49,90 @@ namespace System.Text.JsonLab
             return result;
         }
 
-        internal JsonObject(ReadOnlySpan<byte> values, ReadOnlySpan<byte> db, MemoryPool<byte> pool = null, IMemoryOwner<byte> dbMemory = null)
+        internal JsonObject(ReadOnlySpan<byte> values, Span<byte> db)
         {
             _db = db;
             _values = values;
-            _pool = pool;
-            _dbMemory = dbMemory;
+        }
+
+        public bool TryGetChild(out JsonObject value)
+        {
+            DbRow record = Record;
+
+            if (record.Length == 0)
+            {
+                value = default;
+                return false;
+            }
+            if (record.Type != JsonValueType.Object)
+            {
+                JsonThrowHelper.ThrowInvalidOperationException();
+            }
+
+            record = Read<DbRow>(_db.Slice(DbRow.Size));
+
+            int newStart = DbRow.Size + DbRow.Size;
+            int newEnd = newStart + DbRow.Size;
+
+            record = Read<DbRow>(_db.Slice(newStart));
+
+            if (!record.IsSimpleValue)
+            {
+                newEnd = newEnd + DbRow.Size * record.Length;
+            }
+
+            value = new JsonObject(_values, _db.Slice(newStart, newEnd - newStart));
+            return true;
+        }
+
+        public void Remove(JsonObject obj)
+        {
+            ReadOnlySpan<byte> values = obj._values;
+            Span<byte> db = obj._db;
+
+            DbRow objRecord = Read<DbRow>(db);
+            int reduce = objRecord.Length;
+
+            int idx = _db.IndexOf(db);
+
+            int sliceVal = idx + db.Length;
+
+            if (idx >= DbRow.Size)
+            {
+                reduce++;
+                reduce++;
+                idx -= DbRow.Size;
+            }
+
+            Span<byte> temp = _db.Slice(0, idx);
+
+            for (int i = 0; i < temp.Length; i += DbRow.Size)
+            {
+                DbRow record = Read<DbRow>(temp.Slice(i));
+                if (!record.IsSimpleValue)
+                {
+                    DbRow myObj = Read<DbRow>(_db.Slice(i + record.Length * DbRow.Size));
+                    if (myObj.Location > objRecord.Location)
+                    {
+                        DbRow fixup = new DbRow(record.Type, record.Location, record.Length - reduce);
+                        Write(temp.Slice(i), ref fixup);
+                    }
+                }
+            }
+
+            _db.Slice(sliceVal).CopyTo(_db.Slice(idx));
+            _db = _db.Slice(0, _db.Length - db.Length - DbRow.Size);
         }
 
         public bool TryGetValue(Utf8Span propertyName, out JsonObject value)
         {
-            var record = Record;
+            DbRow record = Record;
 
             if (record.Length == 0) {
-                throw new KeyNotFoundException();
+                JsonThrowHelper.ThrowKeyNotFoundException();
             }
             if (record.Type != JsonValueType.Object) {
-                throw new InvalidOperationException();
+                JsonThrowHelper.ThrowInvalidOperationException();
             }
 
             for (int i = DbRow.Size; i <= _db.Length; i += DbRow.Size) {
@@ -60,19 +145,16 @@ namespace System.Text.JsonLab
 
                 if (new Utf8Span(_values.Slice(record.Location, record.Length)) == propertyName) {
                     int newStart = i + DbRow.Size;
-                    int newEnd = newStart + DbRow.Size;
 
                     record = Read<DbRow>(_db.Slice(newStart));
 
-                    if (!record.IsSimpleValue) {
-                        newEnd = newEnd + DbRow.Size * record.Length;
-                    }
+                    int newEnd = record.IsSimpleValue ? i + 2 * DbRow.Size : i + (record.Length + 2) * DbRow.Size;
 
                     value = new JsonObject(_values, _db.Slice(newStart, newEnd - newStart));
                     return true;
                 }
 
-                var valueType = Read<JsonValueType>(_db.Slice(i + DbRow.Size + 8));
+                JsonValueType valueType = Read<JsonValueType>(_db.Slice(i + DbRow.Size + 8));
                 if (valueType != JsonValueType.Object && valueType != JsonValueType.Array) {
                     i += DbRow.Size;
                 }
@@ -87,14 +169,14 @@ namespace System.Text.JsonLab
             var record = Record;
             
             if (record.Length == 0) {
-                throw new KeyNotFoundException();
+                JsonThrowHelper.ThrowKeyNotFoundException();
             }
 
             if (record.Type != JsonValueType.Object) {
-                throw new InvalidOperationException();
+                JsonThrowHelper.ThrowInvalidOperationException();
             }
 
-            for (int i = DbRow.Size; i <= _db.Length; i += DbRow.Size) {
+            for (int i = DbRow.Size; i < _db.Length; i += DbRow.Size) {
                 record = Read<DbRow>(_db.Slice(i));
 
                 if (!record.IsSimpleValue) {
@@ -133,7 +215,8 @@ namespace System.Text.JsonLab
                 {
                     return value;
                 }
-                throw new KeyNotFoundException();
+                JsonThrowHelper.ThrowKeyNotFoundException();
+                return default;
             }
         }
 
@@ -143,20 +226,23 @@ namespace System.Text.JsonLab
                 {
                     return value;
                 }
-                throw new KeyNotFoundException();
+                JsonThrowHelper.ThrowKeyNotFoundException();
+                return default;
             }
         }
 
         public JsonObject this[int index] {
             get {
-                var record = Record;
+                DbRow record = Record;
 
-                if (index < 0 || index >= record.Length) {
-                    throw new IndexOutOfRangeException();
+                if ((uint)index >= (uint)record.Length) {
+                    JsonThrowHelper.ThrowIndexOutOfRangeException();
+                    return default;
                 }
 
                 if (record.Type != JsonValueType.Array) {
-                    throw new InvalidOperationException();
+                    JsonThrowHelper.ThrowInvalidOperationException();
+                    return default;
                 }
 
                 int counter = 0;
@@ -180,7 +266,8 @@ namespace System.Text.JsonLab
                     counter++;
                 }
 
-                throw new IndexOutOfRangeException();
+                JsonThrowHelper.ThrowIndexOutOfRangeException();
+                return default;
             }
         }
 
@@ -191,7 +278,7 @@ namespace System.Text.JsonLab
                 var record = Record;
                 if (record.Type != JsonValueType.Array)
                 {
-                    throw new InvalidOperationException();
+                    JsonThrowHelper.ThrowInvalidOperationException();
                 }
                 return record.Length; 
             }
@@ -199,15 +286,14 @@ namespace System.Text.JsonLab
 
         public static explicit operator string(JsonObject json)
         {
-            var utf8 = (Utf8Span)json;
-            return utf8.ToString();
+            return ((Utf8Span)json).ToString();
         }
 
         public static explicit operator Utf8Span(JsonObject json)
         {
-            var record = json.Record;
+            DbRow record = json.Record;
             if (!record.IsSimpleValue) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             return new Utf8Span(json._values.Slice(record.Location, record.Length));
@@ -218,7 +304,7 @@ namespace System.Text.JsonLab
             DbRow record = json.Record;
             if (!record.IsSimpleValue)
             {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             return new Utf8String(json._values.Slice(record.Location, record.Length));
@@ -228,18 +314,18 @@ namespace System.Text.JsonLab
         {
             var record = json.Record;
             if (!record.IsSimpleValue) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             if (record.Length < 4 || record.Length > 5) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             var slice = json._values.Slice(record.Location);
 
             if (!Utf8Parser.TryParse(slice, out bool result, out _))
             {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
             return result;
         }
@@ -248,14 +334,14 @@ namespace System.Text.JsonLab
         {
             var record = json.Record;
             if (!record.IsSimpleValue) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             var slice = json._values.Slice(record.Location);
 
             if (!Utf8Parser.TryParse(slice, out int result, out _))
             {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
             return result;
         }
@@ -264,7 +350,7 @@ namespace System.Text.JsonLab
         {
             var record = json.Record;
             if (!record.IsSimpleValue) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             int count = record.Location;
@@ -277,7 +363,7 @@ namespace System.Text.JsonLab
             }
 
             if (nextByte < '0' || nextByte > '9' || count - record.Location >= record.Length) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             int integerPart = 0;
@@ -329,22 +415,31 @@ namespace System.Text.JsonLab
             }
 
             if (count - record.Location > record.Length) {
-                throw new InvalidCastException();
+                JsonThrowHelper.ThrowInvalidCastException();
             }
 
             return isNegative ? result * -1 : result;
 
         }
 
-        internal DbRow Record => Read<DbRow>(_db);
-        public JsonValueType Type => Read<JsonValueType>(_db.Slice(8));
+        private DbRow Record => GetRecord(0);
+        private int Location => GetLocation(0);
+        private int Length => GetLength(0);
 
+        private DbRow GetRecord(int index) => index == 0 ? Read<DbRow>(_db) : Read<DbRow>(_db.Slice(index));
+        private int GetLocation(int index) => index == 0 ? Read<int>(_db) : Read<int>(_db.Slice(index));
+        private int GetLength(int index) => index == 0 ? Read<int>(_db.Slice(4)) : Read<int>(_db.Slice(index + 4));
+        private JsonValueType GetType(int index) => index == 0 ? Read<JsonValueType>(_db.Slice(8)) : Read<JsonValueType>(_db.Slice(index + 8));
+
+        public JsonValueType Type => GetType(0);
+
+        // Do not change the order of the enum values, since IsSimpleValue relies on it.
         public enum JsonValueType : byte
         {
-            String = 0,
-            Number = 1,
-            Object = 2,
-            Array  = 3,
+            Object = 0,
+            Array = 1,
+            String = 2,
+            Number = 3,
             True   = 4,
             False  = 5,
             Null   = 6
@@ -352,11 +447,10 @@ namespace System.Text.JsonLab
 
         public bool HasValue()
         {
-            var record = Record;
+            DbRow record = Record;
 
             if (!record.IsSimpleValue) {
-                if (record.Length == 0) return false;
-                return true;
+                return record.Length != 0;
             } else {
                 if (_values[record.Location - 1] == '"' && _values[record.Location + 4] == '"') {
                     return true;
@@ -368,13 +462,8 @@ namespace System.Text.JsonLab
         public void Dispose()
         {
             //if (_pool == null) throw new InvalidOperationException("only root object can (and should) be disposed.");
-            _db = ReadOnlySpan<byte>.Empty;
+            _db = Span<byte>.Empty;
             _values = ReadOnlySpan<byte>.Empty;
-            if (_dbMemory != null)
-            {
-                _dbMemory.Dispose();
-                _dbMemory = null;
-            }
         }
     }
 }