--- a/src/System.Text.JsonLab/System/Text/Json/JsonReader.cs
+++ b/src/System.Text.JsonLab/System/Text/Json/JsonReader.cs
@@ -3,7 +3,6 @@
 
 using System.Buffers;
 using System.Buffers.Reader;
-using System.Runtime.CompilerServices;
 
 namespace System.Text.JsonLab
 {