… src/System.Text.JsonLab/System/Text/Json/JsonParser.cs
… src/System.Text.JsonLab/System/Text/Json/JsonThrowHelper.cs
… tests/Benchmarks/System.Text.JsonLab/JsonParserPerf.cs
… tests/Benchmarks/System.Text.JsonLab/JsonReaderPerf.cs
… tests/System.Text.JsonLab.Tests/JsonObjectTests.cs
… tests/System.Text.JsonLab.Tests/JsonTestHelper.cs
… tests/System.Text.JsonLab.Tests/Resources/TestJson.Designer.cs
… tests/System.Text.JsonLab.Tests/Resources/TestJson.resx