- Use ThrowHelper
- Fix use of var
- Add JsonTestHelper
- Add JsonObject TryGetChild and Remove APIs

``` ini

BenchmarkDotNet=v0.10.14.683-nightly, OS=Windows 10.0.17713
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-alpha1-20180720-2
  [Host]     : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT
  Job-JGLELY : .NET Core 2.1.2 (CoreCLR 4.6.26628.05, CoreFX 4.6.26629.01), 64bit RyuJIT

IterationCount=5  WarmupCount=3  

```
|                  Method |        TestCase |            Mean |           Error |        StdDev | Scaled | ScaledSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------------------ |---------------- |----------------:|----------------:|--------------:|-------:|---------:|---------:|---------:|---------:|----------:|
|         **ParseNewtonsoft** |       **BasicJson** |      **5,648.6 ns** |       **977.86 ns** |     **254.00 ns** |   **1.00** |     **0.00** |   **1.5564** |        **-** |        **-** |    **6536 B** |
|    ParseNewtonsoftUtf16 |       BasicJson |      5,187.2 ns |       180.93 ns |      46.99 ns |   0.92 |     0.04 |   1.5640 |        - |        - |    6568 B |
| ReaderSystemTextJsonLab |       BasicJson |      3,684.8 ns |        74.47 ns |      19.34 ns |   0.65 |     0.03 |   0.0038 |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |   **BasicLargeNum** |      **6,418.3 ns** |       **424.03 ns** |     **110.14 ns** |   **1.00** |     **0.00** |   **1.7319** |        **-** |        **-** |    **7272 B** |
|    ParseNewtonsoftUtf16 |   BasicLargeNum |      6,614.6 ns |     1,394.21 ns |     362.14 ns |   1.03 |     0.05 |   1.7395 |        - |        - |    7304 B |
| ReaderSystemTextJsonLab |   BasicLargeNum |      2,317.7 ns |       171.33 ns |      44.50 ns |   0.36 |     0.01 |   0.0038 |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |       **BroadTree** |     **81,450.9 ns** |     **5,306.21 ns** |   **1,378.27 ns** |   **1.00** |     **0.00** |  **17.0898** |   **0.1221** |        **-** |   **71960 B** |
|    ParseNewtonsoftUtf16 |       BroadTree |     81,380.5 ns |     9,543.23 ns |   2,478.82 ns |   1.00 |     0.03 |  17.0898 |   0.1221 |        - |   71992 B |
| ReaderSystemTextJsonLab |       BroadTree |     80,873.3 ns |     2,748.24 ns |     713.85 ns |   0.99 |     0.02 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |        **DeepTree** |     **46,065.2 ns** |     **3,505.92 ns** |     **910.65 ns** |   **1.00** |     **0.00** |   **8.9111** |        **-** |        **-** |   **37536 B** |
|    ParseNewtonsoftUtf16 |        DeepTree |     45,659.6 ns |     7,928.30 ns |   2,059.35 ns |   0.99 |     0.04 |   8.9111 |        - |        - |   37568 B |
| ReaderSystemTextJsonLab |        DeepTree |     61,119.8 ns |     1,655.52 ns |     430.01 ns |   1.33 |     0.03 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |     **FullSchema1** |     **19,101.8 ns** |     **3,024.46 ns** |     **785.59 ns** |   **1.00** |     **0.00** |   **3.9978** |        **-** |        **-** |   **16816 B** |
|    ParseNewtonsoftUtf16 |     FullSchema1 |     17,450.8 ns |       631.82 ns |     164.11 ns |   0.91 |     0.03 |   3.9978 |        - |        - |   16848 B |
| ReaderSystemTextJsonLab |     FullSchema1 |      6,695.4 ns |       478.87 ns |     124.38 ns |   0.35 |     0.01 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |     **FullSchema2** |     **24,758.2 ns** |     **2,459.28 ns** |     **638.79 ns** |   **1.00** |     **0.00** |   **4.9744** |        **-** |        **-** |   **20952 B** |
|    ParseNewtonsoftUtf16 |     FullSchema2 |     26,410.7 ns |     4,068.53 ns |   1,056.79 ns |   1.07 |     0.05 |   4.9744 |        - |        - |   20984 B |
| ReaderSystemTextJsonLab |     FullSchema2 |     10,529.0 ns |     1,651.02 ns |     428.85 ns |   0.43 |     0.02 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |      **HelloWorld** |        **983.8 ns** |       **167.49 ns** |      **43.50 ns** |   **1.00** |     **0.00** |   **0.7362** |        **-** |        **-** |    **3096 B** |
|    ParseNewtonsoftUtf16 |      HelloWorld |        926.4 ns |       214.65 ns |      55.75 ns |   0.94 |     0.06 |   0.7439 |        - |        - |    3128 B |
| ReaderSystemTextJsonLab |      HelloWorld |        457.2 ns |        42.90 ns |      11.14 ns |   0.47 |     0.02 |   0.0052 |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |        **Json400B** |      **8,638.7 ns** |     **1,287.46 ns** |     **334.41 ns** |   **1.00** |     **0.00** |   **1.9684** |        **-** |        **-** |    **8304 B** |
|    ParseNewtonsoftUtf16 |        Json400B |      9,002.4 ns |     2,415.30 ns |     627.36 ns |   1.04 |     0.07 |   1.9836 |        - |        - |    8336 B |
| ReaderSystemTextJsonLab |        Json400B |      4,402.8 ns |       559.52 ns |     145.33 ns |   0.51 |     0.02 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |       **Json400KB** | **10,065,638.1 ns** |   **687,598.56 ns** | **178,601.17 ns** |   **1.00** |     **0.00** | **750.0000** | **375.0000** |        **-** | **4721808 B** |
|    ParseNewtonsoftUtf16 |       Json400KB |  9,762,566.3 ns | 1,704,063.59 ns | 442,624.18 ns |   0.97 |     0.04 | 750.0000 | 375.0000 |        - | 4721840 B |
| ReaderSystemTextJsonLab |       Json400KB |  5,533,148.7 ns |   241,622.68 ns |  62,760.59 ns |   0.55 |     0.01 | 328.1250 | 328.1250 | 328.1250 | 1711944 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |        **Json40KB** |    **665,240.4 ns** |    **65,487.13 ns** |  **17,010.04 ns** |   **1.00** |     **0.00** |  **85.9375** |  **35.1563** |        **-** |  **482368 B** |
|    ParseNewtonsoftUtf16 |        Json40KB |    674,611.5 ns |    49,729.94 ns |  12,917.17 ns |   1.01 |     0.03 |  87.8906 |  36.1328 |        - |  482400 B |
| ReaderSystemTextJsonLab |        Json40KB |    454,512.1 ns |    83,537.52 ns |  21,698.56 ns |   0.68 |     0.03 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |         **Json4KB** |     **65,033.6 ns** |    **11,822.73 ns** |   **3,070.91 ns** |   **1.00** |     **0.00** |  **11.8408** |        **-** |        **-** |   **49744 B** |
|    ParseNewtonsoftUtf16 |         Json4KB |     63,758.1 ns |     3,913.43 ns |   1,016.50 ns |   0.98 |     0.04 |  11.8408 |        - |        - |   49776 B |
| ReaderSystemTextJsonLab |         Json4KB |     43,706.7 ns |     9,388.00 ns |   2,438.50 ns |   0.67 |     0.04 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |   **LotsOfNumbers** |     **24,908.5 ns** |     **7,735.46 ns** |   **2,009.26 ns** |   **1.00** |     **0.00** |   **4.9438** |        **-** |        **-** |   **20760 B** |
|    ParseNewtonsoftUtf16 |   LotsOfNumbers |     26,720.1 ns |     5,088.57 ns |   1,321.74 ns |   1.08 |     0.09 |   4.9438 |        - |        - |   20792 B |
| ReaderSystemTextJsonLab |   LotsOfNumbers |     12,557.7 ns |     1,328.99 ns |     345.20 ns |   0.51 |     0.04 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |   **LotsOfStrings** |     **12,019.8 ns** |       **903.95 ns** |     **234.80 ns** |   **1.00** |     **0.00** |   **2.5787** |        **-** |        **-** |   **10832 B** |
|    ParseNewtonsoftUtf16 |   LotsOfStrings |     13,427.3 ns |     6,097.10 ns |   1,583.70 ns |   1.12 |     0.12 |   2.5787 |        - |        - |   10864 B |
| ReaderSystemTextJsonLab |   LotsOfStrings |     11,340.3 ns |       933.54 ns |     242.48 ns |   0.94 |     0.02 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** | **ProjectLockJson** |  **3,324,947.0 ns** |   **442,119.98 ns** | **114,839.02 ns** |   **1.00** |     **0.00** | **312.5000** | **156.2500** |        **-** | **1874584 B** |
|    ParseNewtonsoftUtf16 | ProjectLockJson |  3,186,968.8 ns |   462,143.23 ns | 120,039.99 ns |   0.96 |     0.04 | 312.5000 | 156.2500 |        - | 1874616 B |
| ReaderSystemTextJsonLab | ProjectLockJson |  1,890,295.9 ns |   183,876.16 ns |  47,761.15 ns |   0.57 |     0.02 |        - |        - |        - |      24 B |
|                         |                 |                 |                 |               |        |          |          |          |          |           |
|         **ParseNewtonsoft** |  **SpecialNumForm** |     **13,446.7 ns** |       **394.42 ns** |     **102.45 ns** |   **1.00** |     **0.00** |   **2.3346** |        **-** |        **-** |    **9856 B** |
|    ParseNewtonsoftUtf16 |  SpecialNumForm |     14,391.0 ns |     5,567.54 ns |   1,446.15 ns |   1.07 |     0.10 |   2.3499 |        - |        - |    9888 B |
| ReaderSystemTextJsonLab |  SpecialNumForm |      3,540.3 ns |       195.17 ns |      50.69 ns |   0.26 |     0.00 |   0.0038 |        - |        - |      24 B |